+++ src/controllers/users.ts
@@ -1,41 +1,8 @@
 import { User, Site, Organization, OrganizationPermission, SitePermission } from '../entity';
-import { getManager, In } from 'typeorm';
+import { getManager } from 'typeorm';
 import { NotFoundError } from '../middleware/error/errors';
 import { uniq } from 'underscore';
-
-/**
- * Fills out the organization and site ID fields on a user object
- * by mapping any linked permissions to their respective IDs.
- * SiteIDs are comprehensively generated by supplementing the user's
- * sitePerms with all sites under the purview of organizations the
- * user can also access.
- */
-const mapUserPermsToIds = async (user: User) => {
-
-  // Get all orgs the user has access to, and then use those
-  // to find all sites the user has access to
-  const orgIds = (user.orgPermissions || []).map(
-    (perm) => perm.organizationId
-  );
-  const sitesFromOrgs = orgIds.length
-    ? await getManager().find(Site, {
-        where: { organizationId: In(orgIds) },
-      })
-    : [];
-
-  // Make the collection we got from searching by org distinct
-  const distinctSiteIds = Array.from(
-    new Set([
-      ...(user.sitePermissions || []).map((perm) => perm.siteId),
-      ...sitesFromOrgs.map((site) => site.id),
-    ])
-  );
-
-  // Augment and return the user object
-  user.organizationIds = orgIds;
-  user.siteIds = distinctSiteIds;
-  return user;
-}
+import { mapUserPermsToIds } from '../utils/mapPermsToIds';
 
 /**
  * Augment a user object with site and organization data.
@@ -63,17 +30,6 @@ export const addSiteAndOrgDataToUser = async (user: User) => {
   user.organizations = organizations;
 };
 
-/**
- * TODO: Currently, this function only returns users who have existing
- * organization permissions. Which means the admin view table of users
- * will only show users who already have permissions set. That seems
- * like not what we want to do, right? (Example: a user currently has
- * only site perms, but should have perms for their whole org; this
- * query wouldn't return the user. Example 2: an admin mistakenly
- * forgot to give a user any permissions at all when creating their
- * account. That user won't show up here so that the situation can be
- * fixed.)
- */
 export const getUsers = async (): Promise<User[]> =>
   getManager().query(
     `SELECT
